+++ CHANGELOG.md
@@ -6,8 +6,9 @@ develop
 
 Backward compatibility breaks:
 
-- DBAL extension removed
-- PHPBench Reports extension removed
+- DBAL extension removed.
+- PHPBench Reports extension removed.
+- Removed `--query` featre (only worked with DBAL, too complex).
 
 Features